requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
transformers>=4.30.0
faster-whisper>=0.10.0
pyannote.audio>=3.1.0

# Sentiment Analysis Models
sentencepiece>=0.1.99
protobuf>=3.20.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
sounddevice>=0.4.6
webrtcvad>=2.0.10

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Visualization (for demo)
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
sqlalchemy>=2.0.0

# Optional: For faster inference
accelerate>=0.20.0
optimum>=1.8.0
